requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic==2.11.9
openai==2.0.0
python-dotenv==1.1.1
python-multipart==0.0.20
